pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "img2sound"
version = "0.1.0"
description = "Convert images and video to audio using scanline and spectral sonification"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "img2sound contributors"}
]
keywords = ["sonification", "audio", "image", "video", "glitch", "experimental"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Artistic Software",
]
dependencies = [
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.11.0",
    "opencv-python>=4.8.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
img2sound = "img2sound.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
